Implement piece rotation with some leniency
When rotating a piece next to the walls, it can be "pushed" a little to
the side to accomodate the rotation. The code is a mess, but it works.
It should be rethinked and refactored. Also, when a situation like the
following occurs:

.......
..OOX..
..XOO..
....X..

Where X are consolidated blocks and O is the piece, the piece will pop
out of the blocks when rotated. Maybe this is the expected behavior,
but it's still a little strange.