Implement line clear effect
The lines being cleared blink now, but the code is an even uglier mess.

Seems like a state machine really is the best approach to orchestrate
every element in a game. The main difficultiy, besides deciding how to
best organize and modularize the game, is how to make every element get
along well with each other. how to do things in order, exchange
messages, etc.